4.0

+++ src/main/java/com/infinitus/bms_oa/bms_op/empty/Bms_po_api_return.java
@@ -2,10 +2,13 @@ package com.infinitus.bms_oa.bms_op.empty;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @Data
 public class Bms_po_api_return {
     private String invoiceNO;
-    private Integer OP_ID;
+    private long OP_ID;
     private String Msg_Type;
     private String Msg_Log;
     private String Docno;
